feat: expose `build` and `version` commands from public API
