+++ src/lib/commands/version.command.ts
@@ -1,7 +1,9 @@
-import * as rollup from 'rollup';
+import { VERSION as ROLLUP_VERSION } from 'rollup';
 import { VERSION as COMPILER_VERSION } from '@angular/compiler';
 import { VERSION as COMPILER_CLI_VERSION } from '@angular/compiler-cli';
 import { version as TS_VERSION } from 'typescript';
+import { Command } from './command';
+
 const TSICKLE_VERSION = require('tsickle/package.json').version;
 
 let NG_PACKAGR_VERSION: string = 'unknown';
@@ -21,13 +23,14 @@ function tryReadVersion(paths: string[] = []) {
   }
 }
 
-tryReadVersion(['../../package.json', '../../../package.json']);
+export const version: Command<undefined, void> =
+  () => {
+    tryReadVersion(['../../package.json', '../../../package.json']);
 
-export function printVersionInfo() {
-  console.log(`ng-packagr:            ` + NG_PACKAGR_VERSION);
-  console.log(`@angular/compiler:     ` + COMPILER_VERSION.full);
-  console.log(`@angular/compiler-cli: ` + COMPILER_CLI_VERSION.full);
-  console.log(`rollup:                ` + (rollup as any).VERSION);
-  console.log(`tsickle:               ` + TSICKLE_VERSION);
-  console.log(`typescript:            ` + TS_VERSION);
-}
+    console.log(`ng-packagr:            ` + NG_PACKAGR_VERSION);
+    console.log(`@angular/compiler:     ` + COMPILER_VERSION.full);
+    console.log(`@angular/compiler-cli: ` + COMPILER_CLI_VERSION.full);
+    console.log(`rollup:                ` + ROLLUP_VERSION);
+    console.log(`tsickle:               ` + TSICKLE_VERSION);
+    console.log(`typescript:            ` + TS_VERSION);
+  };
