--- src/lib/domain/cli-arguments.ts
@@ -1,8 +1,0 @@
-/**
- * CLI arguments passed to the `ng-packagr` binary and `createNgPackage()`.
- */
-export interface CliArguments {
-
-  /** Path to the project file 'package.json', 'ng-package.json', or 'ng-package.js'. */
-  project: string
-}
