+++ src/public_api.ts
@@ -1,2 +1,6 @@
-export { CliArguments } from './lib/domain/cli-arguments';
+export * from './lib/commands/command';
+export * from './lib/commands/build.command';
+export * from './lib/commands/version.command';
+
+// XX: backward-compatibility export
 export { createNgPackage as createNgPackage } from './lib/ng-packagr';
